refactor : 진위여부 판별 request, response DTO 패키지 분리해서 저장 #26
