+++ src/main/java/Seoul_Milk/sm_server/domain/taxValidation/dto/response/NonVerifiedTaxValidationResponseDTO.java
@@ -1,4 +1,4 @@
-package Seoul_Milk.sm_server.domain.taxValidation.dto;
+package Seoul_Milk.sm_server.domain.taxValidation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
